@@ -1,58 +1,52 @@
 {
-  "name": "bugsplat-ng4",
-  "version": "2.1.1",
-  "description": "BugSplat error handling for Angular 4",
-  "main": "./src/index.ts",
-  "directories": {
-    "test": "test"
-  },
+  "name": "my-angular6-crasher",
+  "version": "0.0.0",
   "scripts": {
-    "build": "tsc",
-    "test": "karma start karma.conf.js './test/**.spec.ts'",
-    "acceptance": "karma start karma.conf.js './test/**.acceptance-spec.ts'"
+    "ng": "ng",
+    "start": "ng serve",
+    "build": "ng build",
+    "test": "ng test",
+    "lint": "ng lint",
+    "e2e": "ng e2e"
   },
-  "keywords": [
-    "bugsplat",
-    "angular",
-    "4",
-    "error",
-    "handler",
-    "crash",
-    "reporting"
-  ],
-  "homepage": "https://www.bugsplat.com",
-  "bugs": "https://github.com/BugSplat-Git/bugsplat-ng4/issues",
-  "repository": {
-    "type": "git",
-    "url": "https://github.com/BugSplat-Git/bugsplat-ng4.git"
-  },
-  "author": "@bobbyg603",
-  "license": "MIT",
-  "peerDependencies": {
-    "@angular/core": "^4.3.6",
-    "@angular/common": "^4.3.6",
-    "@angular/http": "^4.3.6",
-    "rxjs": "^5.4.3"
+  "private": true,
+  "dependencies": {
+    "@angular/common": "^6.0.3",
+    "@angular/core": "^6.0.3",
+    "@angular/compiler": "^6.0.3",
+    "@angular/animations": "^6.0.3",
+    "@angular/forms": "^6.0.3",
+    "@angular/http": "^6.0.3",
+    "@angular/platform-browser": "^6.0.3",
+    "@angular/platform-browser-dynamic": "^6.0.3",
+    "@angular/router": "^6.0.3",
+    "core-js": "^2.5.4",
+    "rxjs": "^6.0.0",
+    "zone.js": "^0.8.26"
   },
   "devDependencies": {
-    "@angular/common": "^4.3.6",
-    "@angular/compiler": "^4.3.6",
-    "@angular/core": "^4.3.6",
-    "@angular/http": "^4.3.6",
-    "@angular/platform-browser": "^4.3.6",
-    "@angular/platform-browser-dynamic": "^4.3.6",
-    "@types/core-js": "^0.9.43",
-    "@types/jasmine": "^2.5.54",
-    "core-js": "^2.5.1",
-    "jasmine-core": "^2.8.0",
-    "karma": "^1.7.1",
-    "karma-chrome-launcher": "^2.2.0",
-    "karma-firefox-launcher": "^1.0.1",
-    "karma-jasmine": "^1.1.0",
+    "@angular-devkit/build-angular": "~0.6.8",
+    "@angular-devkit/build-ng-packagr": "~0.6.8",
+    "@angular/cli": "~6.0.8",
+    "@angular/compiler-cli": "^6.0.3",
+    "@angular/language-service": "^6.0.3",
+    "@types/jasmine": "~2.8.6",
+    "@types/jasminewd2": "~2.0.3",
+    "@types/node": "~8.9.4",
+    "codelyzer": "~4.2.1",
+    "jasmine-core": "~2.99.1",
+    "jasmine-spec-reporter": "~4.2.1",
+    "karma": "~1.7.1",
+    "karma-chrome-launcher": "~2.2.0",
+    "karma-coverage-istanbul-reporter": "~2.0.0",
+    "karma-jasmine": "~1.1.1",
     "karma-jasmine-html-reporter": "^0.2.2",
-    "karma-typescript": "^3.0.5",
-    "rxjs": "^5.4.3",
-    "typescript": "^2.5.2",
-    "zone.js": "^0.8.17"
+    "ng-packagr": "^3.0.0-rc.2",
+    "protractor": "~5.3.0",
+    "ts-node": "~5.0.1",
+    "tsickle": "^0.32.0",
+    "tslib": "^1.7.1",
+    "tslint": "~5.9.1",
+    "typescript": "~2.7.2"
   }
 }
