@@ -1,10 +1,8 @@
-
 import { BrowserModule } from '@angular/platform-browser';
-import { NgModule, ErrorHandler } from '@angular/core';
+import { NgModule } from '@angular/core';
 import { HttpClientModule } from "@angular/common/http";
 import { BugSplatErrorHandler } from "./bugsplat-error-handler";
 import { BugSplat } from "./bugsplat";
-import { Logger, BugSplatLogger } from "./bugsplat-logger";
 
 @NgModule({
     imports: [
@@ -12,9 +10,7 @@ import { Logger, BugSplatLogger } from "./bugsplat-logger";
         HttpClientModule
     ],
     providers: [
-        BugSplatErrorHandler
-    ],
-    exports: [
+        BugSplatErrorHandler,
         BugSplat
     ]
 })
